Add files via upload
Server app that accepts jpeg, png and mp4 files as inputs. The app then classifies the inputs as 1 of 43 distinct classes of traffic signs based on the trained CNN model and returns this classification to the web client. In the case of mp4 upload, the server app extracts still images from the video file and classifies them.